Refactor prepareExpoURLToClipboard function
